image-analysis-service/requirements.txt
flask==2.0.1
werkzeug==2.0.1
pillow==8.3.1
numpy==1.21.1
opencv-python-headless
pytesseract==0.3.8 
paddleocr
paddlepaddle
requests
numpy
tika

